Refactored Thrive version reader to a separate file
